refactor(imt)!: create two different imt libs
IMT implementations have been separated to allow them to be versioned more precisely.

re #282
