--- packages/imt/src/imt/index.ts
@@ -1,4 +1,0 @@
-import IMT from "./imt"
-
-export { IMT }
-export * from "./types"
