+++ packages/lean-imt/typedoc.json
@@ -1,0 +1,3 @@
+{
+    "entryPoints": ["src/index.ts"]
+}
